alt-f: ipkg: update binary package to work with index script
The legacy ipkg binary was in the debian binary format, which did
not work with our ipkg-make-index script. Since this is the only
package not built from source, we just bump the version and
repackage it using our tools. This was easier than updating the
indexing script to handle it.

The package should now appear in the auto-generated index,
allowing all other Alt-F packages to be installed.